refactor: Community.ts 내용 분리
