+++ src/constants/associationRank.ts
@@ -1,0 +1,6 @@
+export const AssociationRankMapping: { [key: string]: string } = {
+  CHAIRMAN: '회장',
+  EXECUTIVE: '임원진',
+  MEMBER: '회원',
+  none: '회원',
+};
