+++ src/constants/board.ts
@@ -1,0 +1,7 @@
+// parameter로 사용할 boardType
+export const BoardTypeMapping: { [key: string]: string } = {
+  '협회 공지': 'association-notice',
+  '공단 공지': 'service-notice',
+  '정보 공유': 'information-sharing',
+  '참여 신청': 'participation-application',
+};
