+++ src/constants/institutionRank.ts
@@ -1,0 +1,6 @@
+export const InstitutionRankMapping: { [key: string]: string } = {
+  CENTER_DIRECTOR: '회장',
+  REPRESENTATIVE: '임원진',
+  SOCIAL_WORKER: '회원',
+  none: '회원',
+};
